src/requirements.txt
betterlogging~=1.0.0
websockets~=13.1
redis~=5.1.1
marzban~=0.2.9
environs~=11.0.0
SQLAlchemy[asyncio]~=2.0.36
uvicorn~=0.32.0
python-dotenv~=1.0.1
fastapi~=0.115.2
pydantic~=2.9.2
starlette~=0.40.0
httpx~=0.27.2
aiosqlite
alembic~=1.14.0
asyncpg>=0.29.0
paramiko~=3.5.0
psycopg2-binary
argon2_cffi>=23.1.0
itsdangerous>=2.2.0